[Mod] simplify rpc module
